requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
django-cors-headers==4.4.0
django-health-check==3.18.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular-sidecar==2024.7.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==23.0.0
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==2.0.1
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
